% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WheresCroc.R
\name{manualWC}
\alias{manualWC}
\title{manualWC}
\usage{
manualWC(moveInfo, readings, positions, edges, probs)
}
\arguments{
\item{moveInfo}{See runWheresCroc for details}

\item{readings}{See runWheresCroc for details}

\item{positions}{See runWheresCroc for details}

\item{edges}{See runWheresCroc for details}

\item{probs}{See runWheresCroc for details}
}
\value{
See runWheresCroc for details
}
\description{
Control function for Where's Croc that allows manual play using keyboard.
}
